extract strings file to strings_func like in spark grouping
