+++ native/spark-expr/src/string_funcs/mod.rs
@@ -1,0 +1,7 @@
+mod string_space;
+mod prediction;
+mod substring;
+
+pub use prediction::*;
+pub use string_space::StringSpaceExpr;
+pub use substring::SubstringExpr;
